@@ -3,7 +3,6 @@ package com.concough.android.concough;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.NumberPicker;
@@ -14,10 +13,11 @@ import com.concough.android.singletons.FormatterSingleton;
 import com.concough.android.utils.PersianCalendar;
 
 import java.text.ParseException;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 
-public class SignupMoreInfo2Activity extends AppCompatActivity {
+public class SignupMoreInfo2Activity extends TopNavigationActivity {
     private static String TAG = "SignupMoreInfo2Activity";
 
     private NumberPicker numberPicker;
@@ -40,7 +40,7 @@ public class SignupMoreInfo2Activity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_signup_more_info2);
 
-        infoTextView = (TextView)findViewById(R.id.signupInfo2A_infoTextView);
+        infoTextView = (TextView) findViewById(R.id.signupInfo2A_infoTextView);
         numberPicker = (NumberPicker) findViewById(R.id.signupInfo2A_numberPicker);
 
         int jdate = PersianCalendar.getPersianYear(new Date());
@@ -66,27 +66,56 @@ public class SignupMoreInfo2Activity extends AppCompatActivity {
         View.OnClickListener nextListener = new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                selectedYear = Integer.valueOf(numberPicker.getValue());
 
-                Calendar gDate = PersianCalendar.getGregorainCalendar(selectedYear, 1, 1, 1, 1, 1);
                 try {
+                    selectedYear = numberPicker.getValue();
+                    Calendar gDate = PersianCalendar.getGregorainCalendar(selectedYear, 1, 1, 1, 1, 1);
+
                     String s = FormatterSingleton.getInstance().getUTCDateFormatter().format(gDate.getTime());
                     SignupMoreInfo1Activity.signupInfo.setBirthday(FormatterSingleton.getInstance().getUTCShortDateFormatter().parse(s));
+
+                    Intent i = SignupMoreInfo3Activity.newIntent(SignupMoreInfo2Activity.this);
+                    startActivity(i);
                 } catch (ParseException e) {
 
-                }
+            }
 
-                Intent i = SignupMoreInfo3Activity.newIntent(SignupMoreInfo2Activity.this);
-                startActivity(i);
 
                 //Toast.makeText(getApplicationContext(), "Date Picked: " + String.valueOf(selectedYear), Toast.LENGTH_LONG).show();
             }
 
         };
 
-
         nextButton = (Button) findViewById(R.id.signupInfo2A_nextButton);
         nextButton.setTypeface(FontCacheSingleton.getInstance(getApplicationContext()).getBold());
         nextButton.setOnClickListener(nextListener);
+
+        actionBarSet();
     }
+
+    private void actionBarSet() {
+        ArrayList<ButtonDetail> buttonDetailArrayList = new ArrayList<>();
+
+        super.clickEventInterface = new OnClickEventInterface() {
+            @Override
+            public void OnButtonClicked(int id) {
+
+            }
+
+            @Override
+            public void OnBackClicked() {
+                onBackPressed();
+            }
+
+            @Override
+            public void OnTitleClicked() {
+
+            }
+        };
+
+
+        super.createActionBar("کنکوق", true, buttonDetailArrayList);
+    }
+
+
 }
