@@ -11,57 +11,75 @@ class EntranceOpenedCountModelHandler {
         val TAG: String = "EntranceOpenedCountModelHandler"
 
         @JvmStatic
-        fun update(context: Context, entranceUniqueId: String, type: String) : Boolean {
-            val record = EntranceOpenedCountModelHandler.getByType(context, entranceUniqueId, type)
+        fun update(context: Context, username: String, entranceUniqueId: String, type: String) : Boolean {
+            val record = EntranceOpenedCountModelHandler.getByType(context, username, entranceUniqueId, type)
             if (record != null) {
                 try {
-                    RealmSingleton.getInstance(context).DefaultRealm.beginTransaction()
-                    record.count += 1
-                    RealmSingleton.getInstance(context).DefaultRealm.commitTransaction()
+                    RealmSingleton.getInstance(context).DefaultRealm.executeTransaction {
+                        record.count += 1
+                    }
+//                    RealmSingleton.getInstance(context).DefaultRealm.beginTransaction()
+//                    RealmSingleton.getInstance(context).DefaultRealm.commitTransaction()
 
                     return true
-                } catch (exc: Exception) {}
+                } catch (exc: Exception) {
+//                    RealmSingleton.getInstance(context).DefaultRealm.cancelTransaction()
+                }
             } else {
                 val record = EntranceOpenedCountModel()
                 record.entranceUniqueId = entranceUniqueId
                 record.type = type
                 record.count = 1
+                record.username = username
 
                 try {
-                    RealmSingleton.getInstance(context).DefaultRealm.beginTransaction()
-                    RealmSingleton.getInstance(context).DefaultRealm.copyToRealm(record)
-                    RealmSingleton.getInstance(context).DefaultRealm.commitTransaction()
+                    RealmSingleton.getInstance(context).DefaultRealm.executeTransaction {
+                        RealmSingleton.getInstance(context).DefaultRealm.copyToRealm(record)
+                    }
+//                    RealmSingleton.getInstance(context).DefaultRealm.beginTransaction()
+//                    RealmSingleton.getInstance(context).DefaultRealm.commitTransaction()
                     return true
-                } catch (exc: Exception) {}
+                } catch (exc: Exception) {
+//                    RealmSingleton.getInstance(context).DefaultRealm.cancelTransaction()
+                }
             }
 
             return false
         }
 
         @JvmStatic
-        fun getByType(context: Context, entranceUniqueId: String, type: String) : EntranceOpenedCountModel? {
+        fun getByType(context: Context, username: String, entranceUniqueId: String, type: String) : EntranceOpenedCountModel? {
             return RealmSingleton.getInstance(context).DefaultRealm.where(EntranceOpenedCountModel::class.java)
-                    .equalTo("entranceUniqueId", entranceUniqueId).equalTo("type", type).findFirst()
+                    .equalTo("entranceUniqueId", entranceUniqueId).equalTo("type", type)
+                    .equalTo("username", username)
+                    .findFirst()
         }
 
         @JvmStatic
-        fun countByEntranceId(context: Context, entranceUniqueId: String) : Int {
+        fun countByEntranceId(context: Context, username: String, entranceUniqueId: String) : Int {
             val totalCount = RealmSingleton.getInstance(context).DefaultRealm.where(EntranceOpenedCountModel::class.java)
-                    .equalTo("entranceUniqueId", entranceUniqueId).findAll().sum("count")
+                    .equalTo("entranceUniqueId", entranceUniqueId)
+                    .equalTo("username", username)
+                    .findAll().sum("count")
 
             return totalCount?.toInt() ?: 0
 
         }
 
-        fun removeByEntranceId(context: Context, entranceUniqueId: String): Boolean {
+        @JvmStatic
+        fun removeByEntranceId(context: Context, username: String, entranceUniqueId: String): Boolean {
             val opened = RealmSingleton.getInstance(context).DefaultRealm.where(EntranceOpenedCountModel::class.java)
+                    .equalTo("username", username)
                     .equalTo("entranceUniqueId", entranceUniqueId).findAll()
 
             try {
-                RealmSingleton.getInstance(context).DefaultRealm.beginTransaction()
-                opened.deleteAllFromRealm()
-                RealmSingleton.getInstance(context).DefaultRealm.commitTransaction()
+                RealmSingleton.getInstance(context).DefaultRealm.executeTransaction {
+                    opened.deleteAllFromRealm()
+                }
+//                RealmSingleton.getInstance(context).DefaultRealm.beginTransaction()
+//                RealmSingleton.getInstance(context).DefaultRealm.commitTransaction()
             } catch (exc: Exception) {
+//                RealmSingleton.getInstance(context).DefaultRealm.cancelTransaction()
                 return false
             }
 
