@@ -1,6 +1,10 @@
 package com.concough.android.rest
 
 import android.content.Context
+import android.util.Log
+import com.concough.android.settings.CONNECT_TIMEOUT
+import com.concough.android.settings.READ_TIMEOUT
+import com.concough.android.singletons.RetrofitSSLClientSingleton
 import com.concough.android.singletons.TokenHandlerSingleton
 import com.concough.android.singletons.UrlMakerSingleton
 import com.concough.android.structures.HTTPErrorType
@@ -8,12 +12,14 @@ import com.concough.android.structures.NetworkErrorType
 import com.google.gson.Gson
 import com.google.gson.JsonElement
 import com.google.gson.JsonParseException
+import okhttp3.OkHttpClient
 import okhttp3.ResponseBody
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
 
 /**
  * Created by abolfazl on 7/10/17.
@@ -34,10 +40,14 @@ class ActivityRestAPIClass {
             }
 
             TokenHandlerSingleton.getInstance(context).assureAuthorized(completion = {authenticated, error ->
+                Log.d(TAG,"Authenticated : " +authenticated)
+                Log.d(TAG,"Error : " +error)
                 if (authenticated && error == HTTPErrorType.Success) {
                     val headers = TokenHandlerSingleton.getInstance(context).getHeader()
+                    Log.d("HEADER REST:",headers.toString())
 
-                    val Obj = Retrofit.Builder().baseUrl(fullPath).addConverterFactory(GsonConverterFactory.create()).build()
+                    val client = RetrofitSSLClientSingleton.getInstance().getBuilder().build()
+                    val Obj = Retrofit.Builder().baseUrl(fullPath).client(client).addConverterFactory(GsonConverterFactory.create()).build()
                     val profile = Obj.create(RestAPIService::class.java)
                     val request = profile.get(url = fullPath, headers = headers!!)
 
@@ -61,12 +71,12 @@ class ActivityRestAPIClass {
                                     }
                                 }
                                 HTTPErrorType.UnAuthorized, HTTPErrorType.ForbiddenAccess -> {
-                                    TokenHandlerSingleton.getInstance(context).assureAuthorized(true, completion = {authenticated, error ->
-                                        if (authenticated && error == HTTPErrorType.Success) {
+                                    TokenHandlerSingleton.getInstance(context).assureAuthorized(true, completion = {authenticated1, error1 ->
+                                        if (authenticated1 && error1 == HTTPErrorType.Success) {
                                             completion(false, null, HTTPErrorType.Refresh)
                                         }
-                                    }, failure = { error ->
-                                        failure(error)
+                                    }, failure = { error1 ->
+                                        failure(error1)
                                     })
                                 }
                                 else -> completion(false, null, resCode)
