@@ -3,17 +3,19 @@ package com.concough.android.singletons
 import android.content.Context
 import android.content.SharedPreferences
 import com.concough.android.structures.ProfileStruct
-import com.concough.android.utils.KeyChainAccessProxy
 import java.util.*
 
 /**
  * Created by abolfazl on 7/5/17.
  */
 class UserDefaultsSingleton {
+    data class WalletStruct(var cash: Int, var updated: Date)
 
     private var prefs: SharedPreferences
+    private var context: Context
 
     private constructor(context: Context) {
+        this.context = context
         this.prefs = context.getSharedPreferences(FILENAME, Context.MODE_PRIVATE)
     }
 
@@ -21,10 +23,10 @@ class UserDefaultsSingleton {
         val TAG: String = "UserDefaultsSingleton"
         val FILENAME = "profile"
 
-        private var sharedInstance : UserDefaultsSingleton? = null
+        private var sharedInstance: UserDefaultsSingleton? = null
 
         @JvmStatic
-        fun  getInstance(context: Context): UserDefaultsSingleton {
+        fun getInstance(context: Context): UserDefaultsSingleton {
             if (sharedInstance == null)
                 sharedInstance = UserDefaultsSingleton(context)
 
@@ -48,6 +50,13 @@ class UserDefaultsSingleton {
         return this.prefs.getBoolean(key, false)
     }
 
+    private fun setValueAsInt(key: String, value: Int) {
+        this.prefs.edit().putInt(key, value).apply()
+    }
+
+    private fun getValueAsInt(key: String): Int? {
+        return this.prefs.getInt(key, 0)
+    }
 
     public fun clearAll(): Boolean {
         this.prefs.edit().clear().apply()
@@ -61,10 +70,11 @@ class UserDefaultsSingleton {
         return false
     }
 
-    public fun createProfile(firstname: String, lastname: String, grade: String, gender: String, birthday: Date, modified: Date) {
+    public fun createProfile(firstname: String, lastname: String, grade: String, gradeString: String, gender: String, birthday: Date, modified: Date) {
         this.setValueAsString("Profile.Firstname", firstname)
         this.setValueAsString("Profile.Lastname", lastname)
         this.setValueAsString("Profile.Grade", grade)
+        this.setValueAsString("Profile.GradeString", gradeString)
         this.setValueAsString("Profile.Gender", gender)
         this.setValueAsString("Profile.Birthday", FormatterSingleton.getInstance().UTCDateFormatter.format(birthday))
         this.setValueAsString("Profile.Modified", FormatterSingleton.getInstance().UTCDateFormatter.format(modified))
@@ -76,13 +86,14 @@ class UserDefaultsSingleton {
         this.setValueAsString("Profile.Modified", FormatterSingleton.getInstance().UTCDateFormatter.format(modified))
     }
 
-    public fun updateGrade(grade: String, modified: Date) {
+    public fun updateGrade(grade: String, gradeString: String, modified: Date) {
         this.setValueAsString("Profile.Grade", grade)
+        this.setValueAsString("Profile.GradeString", gradeString)
         this.updateModified(modified)
     }
 
-    public fun getUsername(context: Context): String? {
-        return TokenHandlerSingleton.getInstance(context).getUsername()
+    public fun getUsername(): String? {
+        return TokenHandlerSingleton.getInstance(this.context).getUsername()
     }
 
     public fun checkPassword(context: Context, password: String): Boolean {
@@ -96,15 +107,48 @@ class UserDefaultsSingleton {
         val firstname = this.getValueAsString("Profile.Firstname")
         val lastname = this.getValueAsString("Profile.Lastname")
         val grade = this.getValueAsString("Profile.Grade")
+        val gradeString = this.getValueAsString("Profile.GradeString")
         val gender = this.getValueAsString("Profile.Gender")
         val birthday = this.getValueAsString("Profile.Birthday")
         val modified = this.getValueAsString("Profile.Modified")
 
         if (this.prefs.contains("Profile.Created") && this.getValueAsBoolean("Profile.Created")) {
             try {
-                return ProfileStruct(firstname!!, lastname!!, grade!!, gender!!, FormatterSingleton.getInstance().UTCDateFormatter.parse(birthday), FormatterSingleton.getInstance().UTCDateFormatter.parse(modified))
-            } catch (exc: Exception) {}
+                return ProfileStruct(firstname!!, lastname!!, grade!!, gradeString!!, gender!!, FormatterSingleton.getInstance().UTCDateFormatter.parse(birthday), FormatterSingleton.getInstance().UTCDateFormatter.parse(modified))
+            } catch (exc: Exception) {
+            }
         }
         return null
+    }
+
+    fun clearWallet() {
+        this.prefs.edit().remove("Wallet.Created").apply()
+        this.prefs.edit().remove("Wallet.Cash").apply()
+        this.prefs.edit().remove("Wallet.Updated").apply()
+    }
+
+    fun hasWallet(): Boolean {
+        if (this.prefs.contains("Wallet.Created")) {
+            return this.getValueAsBoolean("Wallet.Created")
+        }
+        return false
+    }
+
+    fun getWalletInfo(): WalletStruct? {
+        val cash = this.getValueAsInt("Wallet.Cash")
+        val updated = this.getValueAsString("Wallet.Updated")
+
+        if (cash != null && updated != null) {
+            val up = FormatterSingleton.getInstance().UTCDateFormatter.parse(updated)
+            return WalletStruct(cash, up)
+        }
+
+        return null
+    }
+
+    fun setWalletInfo(cash: Int, updated: String) {
+        this.setValueAsInt("Wallet.Cash", cash)
+        this.setValueAsString("Wallet.Updated", updated)
+        this.setValueAsBoolean("Wallet.Created", true)
     }
 }
