@@ -5,4 +5,4 @@ import java.util.*
 /**
  * Created by abolfazl on 7/5/17.
  */
-data class ProfileStruct(val Firstname: String, val Lastname: String, val Grade: String, val Gender: String, val Birthday: Date, val Modified: Date)
+data class ProfileStruct(val Firstname: String, val Lastname: String, val Grade: String , val GradeString: String, val Gender: String, val Birthday: Date, val Modified: Date)
